TMS-214: Remove solr code, fix test fixtures, tests pass now
