+++ setup.py
@@ -23,14 +23,12 @@ setup(
         'zc.iso8601',
         'zeit.cms >= 3.12.0.dev0',
         'zeit.content.image',
-        'zeit.solr>=2.8.0.dev0',
     ],
     entry_points={
         'fanstatic.libraries': [
             'zeit_find=zeit.find.browser.resources:lib',
         ],
         'console_scripts': [
-            'search-solr=zeit.find.cli:search_solr',
             'search-elastic=zeit.find.cli:search_elastic',
         ],
     },
