+++ src/zeit/find/browser/README.txt
@@ -139,13 +139,13 @@ Search
 The search view returns all data for rendering the result:
 
 >>> import zeit.find.testing
->>> zeit.find.testing.LAYER.set_result('zeit.find', 'tests/data/obama.json')
->>> browser.open('/++skin++cms/search_result?fulltext=Obama')
+>>> zeit.find.testing.LAYER.set_result('zeit.find.tests', 'data/obama.json')
+>>> browser.open('/++skin++cms/search_result?title=Obama')
 >>> result = json.loads(browser.contents)
 >>> pprint.pprint(result)
 {u'results': [{u'application_url': u'http://localhost:8080/++skin++cms',
               ...
-              u'graphical_preview_url': u'http://localhost:8080/++skin++cms/repository/2006/DSC00109_2.JPG/@@thumbnail',
+              u'graphical_preview_url': u'http://localhost:8080/++skin++cms/repository/.../thumbnail',
              ...
  u'template_url': u'http://localhost:8080/++skin++cms/++noop++a12dffa9629480a5cafd9df8a674891e/fanstatic/zeit.find/search_result.jsont'}
