+++ src/zeit/find/configure.zcml
@@ -2,10 +2,9 @@
   xmlns="http://namespaces.zope.org/zope"
   xmlns:grok="http://namespaces.zope.org/grok">
 
-  <utility factory=".elastic.from_product_config" />
-
   <include package="grokcore.component" file="meta.zcml" />
   <grok:grok package="." />
 
-</configure>
+  <utility factory=".search.from_product_config" />
 
+</configure>
